project2-oidc-agent/requirements.txt
authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0